pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ashborn-agent"
version = "0.1.0"
description = "A powerful, cinematic AI terminal agent."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "textual>=0.80.0",
    "python-dotenv",
    "rich",
    "openai",
]

[project.scripts]
ashborn = "ashborn.main:main"

[tool.setuptools]
packages = ["ashborn", "ashborn.cli", "ashborn.tools"]

[tool.setuptools.package-data]
"ashborn" = ["*.py", "*.md", "ashborn-bg.png"]
"ashborn.cli" = ["*.py"]
"ashborn.tools" = ["*.py"]
